Index works async when saved
Refactors our existing async indexing techniques used in work versions
and collections, and puts them into an Indexing module that is also
added to works. Now, these resources are indexed whenever the object is
saved to the database. This should prevent unexpected indexing
discrepancies, but it has the potential to cause cascading index jobs,
where the work will be re-indexed twice whenever one of its versions is.
However, after testing this in a deployed environment, it doesn't seem
to happen during the course of typical operations.
